Completed example1 and example2
`example1` is a simple example of the functionality of Lucene. Two VERY
simple documents were added to Lucene's IndexWriter to keep in cache and
be able to index. Then, an extremely simple query was made to select
text that somewhat contains the query text. It is a simple example of
how Lucene is able to search documents.

`example` is a more concrete example of how I would use Lucene in a
real-world example. I used a Harry Potter books title dataset that
contains Titles, ISBN, among MANY other fields. I added all of the data
set into Lucene's cache and searched the data with, again, a very simple
query.

I am still unsure of how Lucene processes queries, so I will need to
look into that later.